:recycle: Refactor getCorrelationByProcessInstance tests to new specs
