+++ _integration_tests/test/empty_activities/finish_empty_activity.js
